simple_bpsk_SNR_qf 	FIXES: setting value via ControlPort 	changes type of d_WINDOW_SIZE from size_t to int 		the python implementation of gr-ctrlport-monitor can not 		handle size_t. It also recognizes the c++ int as long, so 		it had to be adopted.

+++ include/ccsds/simple_bpsk_SNR_qf.h
@@ -51,8 +51,8 @@ namespace gr {
       virtual float SNR_real() const = 0;
       virtual float SNR_imag() const = 0;
       virtual float SNR_magn() const = 0;
-      virtual size_t window_size() const = 0;
-      virtual void set_window_size(size_t) = 0;
+      virtual int window_size() const = 0;
+      virtual void set_window_size(int) = 0;
     };
 
   } // namespace ccsds
